# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SEN Migration Dashboard API
  description: |
    REST API for the SEN Migration Platform.
    Provides migration status, object progress, rule analysis,
    reconciliation, test scenario management, forensic extraction,
    process mining, data export, and migration planning.
  version: 1.0.0
  contact:
    name: SEN Team
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:4004
    description: Local development
  - url: https://{host}
    description: Production
    variables:
      host:
        default: sapconnect.example.com
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/process-mining/processes/{id}:
    get:
      tags: [ProcessMining]
      summary: Get process config
      operationId: getProcessConfig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Process configuration
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Process not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/process-mining/processes/{id}/reference-model:
    get:
      tags: [ProcessMining]
      summary: Get reference model
      operationId: getProcessReferenceModel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reference process model
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Reference model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/process-mining/analyze:
    post:
      tags: [ProcessMining]
      summary: Full process analysis
      operationId: analyzeProcess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventLogInput'
      responses:
        '200':
          description: Analysis report
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/process-mining/discover:
    post:
      tags: [ProcessMining]
      summary: Heuristic mining
      operationId: discoverProcess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventLogInput'
      responses:
        '200':
          description: Discovered process model
          content:
            application/json:
              schema:
                type: object
  /api/process-mining/conformance:
    post:
      tags: [ProcessMining]
      summary: Conformance checking
      operationId: checkConformance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventLogInput'
      responses:
        '200':
          description: Conformance result
          content:
            application/json:
              schema:
                type: object
  /api/process-mining/performance:
    post:
      tags: [ProcessMining]
      summary: Performance analysis
      operationId: analyzePerformance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventLogInput'
      responses:
        '200':
          description: Performance metrics
          content:
            application/json:
              schema:
                type: object
  /api/process-mining/variants:
    post:
      tags: [ProcessMining]
      summary: Variant analysis
      operationId: analyzeVariants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventLogInput'
      responses:
        '200':
          description: Variant analysis
          content:
            application/json:
              schema:
                type: object
  /api/process-mining/social-network:
    post:
      tags: [ProcessMining]
      summary: Social network mining
      operationId: analyzeSocialNetwork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventLogInput'
      responses:
        '200':
          description: Social network analysis
          content:
            application/json:
              schema:
                type: object
  /api/process-mining/kpis:
    post:
      tags: [ProcessMining]
      summary: KPI calculation
      operationId: calculateKPIs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventLogInput'
      responses:
        '200':
          description: KPI results
          content:
            application/json:
              schema:
                type: object
  /api/process-mining/demo/{processId}:
    get:
      tags: [ProcessMining]
      summary: Demo analysis
      operationId: getDemoAnalysis
      parameters:
        - name: processId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Demo analysis with synthetic data
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Process not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Platform Summary ────────────────────────────────────────
...
  /api/migration/plan:
    post:
      tags: [MigrationPlan]
      summary: Generate migration plan
      operationId: createMigrationPlan
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                forensicResult:
                  type: object
                  description: Forensic extraction results (optional — falls back to server state)
                options:
                  type: object
                  description: Plan generation options
      responses:
        '200':
          description: Generated migration plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPlan'
        '400':
          description: No forensic data available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/migration/plan/latest:
    get:
      tags: [MigrationPlan]
      summary: Get latest migration plan
      operationId: getLatestMigrationPlan
      responses:
        '200':
          description: Latest migration plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPlan'
        '404':
          description: No plan available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Export endpoints ────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Required when API_KEY environment variable is set.
        Pass via X-API-Key header or apiKey query parameter.
        Exempt paths: /health, /ready, /metrics
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    EventLogInput:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            type: object
            required:
              - caseId
              - activity
              - timestamp
            properties:
              caseId:
                type: string
              activity:
                type: string
              timestamp:
                type: string
                format: date-time
              resource:
                type: string
        options:
          type: object
    MigrationPlan:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
        scope:
          type: object
          properties:
            activeModules:
              type: array
              items:
                type: string
            totalObjects:
              type: integer
            totalEstimatedRecords:
              type: integer
        objects:
          type: array
          items:
            type: object
        executionPlan:
          type: object
          properties:
            waves:
              type: array
              items:
                type: object
            totalWaves:
              type: integer
        effort:
          type: object
        risks:
          type: array
          items:
            type: object
        recommendations:
          type: array
          items:
            type: object